restructure code to include data fetching with routes
